+++ app/src/main/java/de/passwordvault/view/entries/activity_entry/EntryViewModel.java
@@ -20,12 +20,18 @@ public class EntryViewModel extends ViewModel {
     @Nullable
     private EntryExtended entry;
 
+    /**
+     * Attribute indicates whether the invisible details are expanded.
+     */
+    private boolean invisibleDetailsExpanded;
+
 
     /**
      * Constructor instantiates a new view model.
      */
     public EntryViewModel() {
         entry = null;
+        invisibleDetailsExpanded = false;
     }
 
 
@@ -47,6 +53,24 @@ public class EntryViewModel extends ViewModel {
      */
     public void setEntry(@NonNull EntryExtended entry) {
         this.entry = entry;
+    }
+
+    /**
+     * Method returns whether the list of invisible details is expanded.
+     *
+     * @return  Whether the list of invisible details is expanded.
+     */
+    public boolean areInvisibleDetailsExpanded() {
+        return invisibleDetailsExpanded;
+    }
+
+    /**
+     * Method changes whether the list of invisible details is expanded.
+     *
+     * @param invisibleDetailsExpanded  Whether the list of invisible details is expanded.
+     */
+    public void setInvisibleDetailsExpanded(boolean invisibleDetailsExpanded) {
+        this.invisibleDetailsExpanded = invisibleDetailsExpanded;
     }
 
 }